requirements.txt
pandas==2.0.3
numpy==1.26.4
scikit-learn==1.4.2
streamlit==1.30.0
pyyaml==6.0.1